Fun_Projects/Quotes/requirements.txt
bcrypt==3.1.4
cffi==1.11.4
Django==1.11.10
pycparser==2.18
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
